Move getPageViews() to SuggestedEditsUserStats
